Documentation and Method Adjustments
Added Doc comments to the Helper Methods in Dungeon. Added comments in User, Dungeon, and Sprite, to indicate where the Getters and Setters are. Renamed the currentDungeonDate() to getDungeonDateByDate(), and adjusted the method accordingly. Adjusted isCompleted() to account for the Status of the DungeonDates. Added the param "Date day" to succeedDay() and failDay(), to allow for setting the status of those specific days. I have not fully implemented that.